Update email_the_data.py
March 25, 2020 - cleaned up code at code at the bottom where we send the email. Created a try, except for sending the email and added a comment concerning gmail. In order to make the code work the gmail account that you are sending the model from must be set as Less Secure app access.  Peter Koppelman